moved gallery view to widgets directory, changed from listview to gridview
